Setup for next release, which shall be 0.1.0.
Awkwardly, I moved the location of Changelog.md after I made the 0.0.7
release to PyPI.  It's conceivable the build tools would have picked
it up in the new location but not in the old.  So its relocation
should be considered part of the 0.1.0 release.

It is not a user visible change anyway, unless browsing on github.